move download image code to a mixin
